refactor: 웹소켓 훅 리팩토링
